Module metatensor::c_api

Rust definition corresponding to metatensor-core C-API.

This module is exported for advanced users of the metatensor crate, but should not be needed by most.

Structs

mts_array_t

mts_block_t

mts_labels_t

mts_sample_mapping_t

mts_tensormap_t

Constants

MTS_BUFFER_SIZE_ERROR

MTS_INTERNAL_ERROR

MTS_INVALID_PARAMETER_ERROR

MTS_IO_ERROR

MTS_SERIALIZATION_ERROR

MTS_SUCCESS

Functions

mts_block⚠

mts_block_add_gradient⚠

mts_block_copy⚠

mts_block_data⚠

mts_block_free⚠

mts_block_gradient⚠

mts_block_gradients_list⚠

mts_block_labels⚠

mts_disable_panic_printing⚠

mts_get_data_origin⚠

mts_labels_clone⚠

mts_labels_create⚠

mts_labels_free⚠

mts_labels_intersection⚠

mts_labels_load⚠

mts_labels_load_buffer⚠

mts_labels_position⚠

mts_labels_save⚠

mts_labels_save_buffer⚠

mts_labels_set_user_data⚠

mts_labels_union⚠

mts_labels_user_data⚠

mts_last_error⚠

mts_register_data_origin⚠

mts_tensormap⚠

mts_tensormap_block_by_id⚠

mts_tensormap_blocks_matching⚠

mts_tensormap_components_to_properties⚠

mts_tensormap_copy⚠

mts_tensormap_free⚠

mts_tensormap_keys⚠

mts_tensormap_keys_to_properties⚠

mts_tensormap_keys_to_samples⚠

mts_tensormap_load⚠

mts_tensormap_load_buffer⚠

mts_tensormap_save⚠

mts_tensormap_save_buffer⚠

mts_version⚠

Type Aliases

mts_create_array_callback_t

mts_data_origin_t

mts_realloc_buffer_t

mts_status_t