metatensor

Module data

Structs

ArrayRef

Reference to a data array in metatensor-core

ArrayRefMut

Mutable reference to a data array in metatensor-core

DLDataType

The data type the tensor can hold. The data type is assumed to follow the native endian-ness. An explicit error message should be raised when attempting to export an array with non-native endianness

DLDevice

A Device for Tensor and operator.

DLPackTensor

A managed DLPack tensor, carrying ownership of the data.

DLPackVersion

The DLPack version.

EmptyArray

An implementation of the Array trait without any data.

MtsArray

Wrapper around mts_array_t that provides a more convenient API to use it in Rust code, and in particular to access the underlying array as an &dyn Any instance where possible.

mts_data_movement_t

Enums

DLDataTypeCode

The type code options DLDataType.

DLDeviceType

The device type in DLDevice.

Traits

Array

The Array trait is used by metatensor to manage different kind of data array with a single API. Metatensor only knows about Box<dyn Array>, and manipulate the data through the functions on this trait.