metatensor
Module errors

Structs

Error

Error type used in metatensor

Constants

LAST_RUST_ERROR

Storage for the last error coming from a Rust function

Functions

check_ptr

Check a pointer allocated by metatensor-core, returning an error if is null

check_status

Check an mts_status_t, returning an error if is it not MTS_SUCCESS