Module metatensor::io

Input/Output facilities for storing crate::TensorMap and crate::Labels on disk

Functions

load

Load the serialized tensor map from the given path.

load_buffer

Load a serialized TensorMap from a buffer.

load_labels

Save labels to the file at the given path.

load_labels_buffer

Load a serialized TensorMap from a buffer.

save

Save the given tensor to a file.

save_buffer

Save the given tensor to an in-memory buffer.

save_labels

Save the given Labels to a file.

save_labels_buffer

Save the given labels to an in-memory buffer.