metatensor
Struct TensorBlockRef

pub struct TensorBlockRef<'a> { /* private fields */ }

Reference to a TensorBlock

Implementations

impl<'a> TensorBlockRef<'a>

pub fn data(&'a self) -> TensorBlockData<'a>

Get all the data and metadata inside this TensorBlockRef as a struct with separate fields, to allow borrowing them separately.

pub fn values(&self) -> ArrayRef<'a>

Get the array for the values in this block

pub fn samples(&self) -> Labels

Get the samples for this block

pub fn components(&self) -> Vec<Labels>

Get the components for this block

pub fn properties(&self) -> Labels

Get the properties for this block

pub fn gradient_list(&self) -> Vec<&'a str>

Get the full list of gradients in this block

pub fn gradient(&self, parameter: &str) -> Option<TensorBlockRef<'a>>

Get the data and metadata for the gradient with respect to the given parameter in this block, if it exists.

pub fn try_clone(&self) -> Result<TensorBlock, Error>

Clone this block, cloning all the data and metadata contained inside.

This can fail if the external data held inside an mts_array_t can not be cloned.

pub fn gradients(&self) -> GradientsIter<'_>

Get an iterator over parameter/TensorBlockRef pairs for all gradients in this block

pub fn save(&self, path: impl AsRef<Path>) -> Result<(), Error>

Save the given block to the file at path

This is a convenience function calling crate::io::save_block

pub fn save_buffer(&self, buffer: &mut Vec<u8>) -> Result<(), Error>

Save the given block to an in-memory buffer

This is a convenience function calling crate::io::save_block_buffer