metatensor::data

Struct DLPackTensor

pub struct DLPackTensor { /* private fields */ }

A managed DLPack tensor, carrying ownership of the data.

Convertion from and to other array types is handled though the different TryFrom implementations.

Implementations

impl DLPackTensor

pub unsafe fn from_ptr(tensor: *mut DLManagedTensorVersioned) -> DLPackTensor

Create a DLPackTensor from a pointer to DLManagedTensorVersioned.

Safety

The DLManagedTensorVersioned should have a valid deleter that can be called from Rust, or have the deleter set to None.

pub unsafe fn from_raw( tensor: NonNull<DLManagedTensorVersioned>, ) -> DLPackTensor

Create a DLPackTensor from a non-null pointer to DLManagedTensorVersioned.

Safety

The same safety requirements as from_ptr apply.

pub fn as_ref(&self) -> DLPackTensorRef<'_>

Get a DLPack tensor reference from this owned tensor

pub fn as_mut(&mut self) -> DLPackTensorRefMut<'_>

Get a mutable DLPack tensor reference from this owned tensor

pub fn version(&self) -> DLPackVersion

Get the ABI version of this DLPack tensor.

pub fn data_ptr<T>(&self) -> Result<*const T, CastError>
where T: DLPackPointerCast,

Get a pointer to data in this tensor. This pointer can be a device pointer according to DLPackTensor::device.

pub fn data_ptr_mut<T>(&mut self) -> Result<*mut T, CastError>
where T: DLPackPointerCast,

Get a mutable pointer to data in this tensor. This pointer can be a device pointer according to DLPackTensor::device.

pub fn device(&self) -> DLDevice

Get the device where the data of this tensor lives.

pub fn dtype(&self) -> DLDataType

Get the data type of this tensor

pub fn n_dims(&self) -> usize

Get the number of dimensions of this tensor

pub fn shape(&self) -> &[i64]

Get the shape of this tensor

pub fn strides(&self) -> Option<&[i64]>

Get the strides of this tensor, if any

pub fn flags(&self) -> u64

Get the raw flags bitfield.

pub fn is_read_only(&self) -> bool

Check if the tensor is explicitly marked as read-only.

pub fn is_copied(&self) -> bool

Check if the tensor is unique/owned (IS_COPIED bit).

pub fn is_subbyte_type_padded(&self) -> bool

Check if the sub-byte types (fp4, fp6) are padded to the next byte, or packed together.

pub fn byte_offset(&self) -> usize

Get the byte offset of this tensor, i.e. how many bytes should be added to DLPackTensor::data_ptr and DLPackTensor::data_ptr_mut to get the first element of the tensor.

pub fn as_dltensor(&self) -> &DLTensor

Get a reference to the underlying DLTensor

pub fn into_raw(self) -> NonNull<DLManagedTensorVersioned>

Consumes the DLPackTensor, returning the underlying raw pointer.

Safety

The caller is responsible for managing the memory and calling the deleter when the tensor is no longer needed.