metatensor

Module io

Input/Output facilities for storing crate::TensorMap and crate::Labels on disk

Functions

load

Load the serialized tensor map from the given path.

load_block

Load previously saved TensorBlock from the file at the given path.

load_block_buffer

Load a serialized TensorBlock from a buffer.

load_buffer

Load a serialized TensorMap from a buffer.

load_labels

Load previously saved Labels from the file at the given path.

load_labels_buffer

Load previously saved Labels from an in-memory buffer.

save

Save the given tensor to a file.

save_block

Save the given block to a file.

save_block_buffer

Save the given block to an in-memory buffer.

save_buffer

Save the given tensor to an in-memory buffer.

save_labels

Save the given Labels to a file.

save_labels_buffer

Save the given labels to an in-memory buffer.